requirements.txt
einops==0.4.1
hydra-core==1.2.0
mne==1.0.3
mne_bids==0.10
natsort==8.2.0
numpy==1.22.0
omegaconf==2.2.3
pandas==1.4.3
scikit_learn==1.1.3
scipy==1.8.1
termcolor==1.1.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
tqdm==4.64.0
transformers==4.24.0
seaborn
mamba-ssm
wandb